[refactor] move generator/discriminator scaffolding to gan_graph
